browser/webpage: Move load_history to browser/webview
This commit moves load_history out of browser/webpage into
browser/webview cleaning up unneeded imports and changing misc/sessions
to call the new method.
